requirements/base.txt
mysqlclient==1.3.7
Django==1.11.1
djangorestframework==3.6.3
django-rest-swagger==2.1.2
markdown==2.6.8
django-filter==1.0.4
django-admin-view-permission==0.6
